+++ public/index.html
@@ -27,7 +27,6 @@
     <title>GlideIM</title>
     <style>
         #root {
-            background-image: url("bg.jpg");
             background-position: center;
         }
     </style>
