+++ src/component/MainPanel.tsx
@@ -12,10 +12,10 @@ import {grey} from "@mui/material/colors";
 import React from "react";
 import {Redirect, Route, RouteComponentProps, Switch, useRouteMatch, withRouter} from "react-router-dom";
 import {Account} from "../im/account";
-import {Chat} from "./chat/Chat";
+import {Chat} from "./session/Chat";
 import {ContactsList} from "./friends/ContactsList";
 import {Square} from "./square/Square";
-import {SessionListView} from "./chat/SessionListView";
+import {SessionListView} from "./session/SessionListView";
 import {ManageAccountsOutlined, MessageOutlined} from "@mui/icons-material";
 import {ChatRoomContainerMobile} from "./chat/ChatRoom";
 import { Profile } from "./Profile";
