+++ src/component/session/SessionListItem.tsx
