--- src/component/version.tsx
@@ -1,10 +1,0 @@
-import {Box} from "@mui/material"
-import {useEffect, useState} from "react"
-
-export function Version() {
-
-
-    return <Box>
-        {/*<Typography variant="caption">build: {version}</Typography>*/}
-    </Box>
-}
